moves addMultiplyNode callback to MultiplyNode.tsx
